[nanocloud 0.7.*] File push support for tunneller - check before rename to keep behavour consistent between OSes
git-svn-id: https://gridkit.googlecode.com/svn/grid-lab/branches/vicluster-0.7.5@3085 4e96dbfe-9890-11de-ac5d-73b9d092dc0a

+++ telecontrol/src/main/java/org/gridkit/vicluster/telecontrol/bootstraper/Tunneller.java
@@ -507,7 +507,9 @@ public class Tunneller extends TunnellerIO {
 				}
 				close(fos);
 				close(in);
-				tempFile.renameTo(targetFile);
+				if (!targetFile.exists()) {
+					tempFile.renameTo(targetFile);
+				}
 				if (tempFile.exists()) {
 					tempFile.delete();
 				}
